Type cast

+++ src/mgd77/mgd77convert_func.c
@@ -1,5 +1,5 @@
 /*--------------------------------------------------------------------
- *	$Id: mgd77convert_func.c,v 1.12 2011-06-22 01:35:00 guru Exp $
+ *	$Id: mgd77convert_func.c,v 1.13 2011-06-22 08:04:02 jluis Exp $
  *
  *    Copyright (c) 2005-2011 by P. Wessel
  *    See README file for copying and redistribution conditions.
@@ -199,9 +199,9 @@ GMT_LONG GMT_mgd77convert (struct GMTAPI_CTRL *API, struct GMT_OPTION *options)
 	/* Parse the command-line arguments */
 
 	GMT = GMT_begin_module (API, "GMT_mgd77convert", &GMT_cpy);		/* Save current state */
-	if ((error = GMT_Parse_Common (API, "-V", "", options))) Return (error);
+	if ((error = GMT_Parse_Common (API, "-V", "", options))) Return ((int)error);
 	Ctrl = (struct MGD77CONVERT_CTRL *) New_mgd77convert_Ctrl (GMT);	/* Allocate and initialize a new control structure */
-	if ((error = GMT_mgd77convert_parse (API, Ctrl, options))) Return (error);
+	if ((error = GMT_mgd77convert_parse (API, Ctrl, options))) Return ((int)error);
 	
 	/*---------------------------- This is the mgd77convert main code ----------------------------*/
